% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{get_anno_storage}
\alias{get_anno_storage}
\alias{clear_anno_storage}
\title{Get annotation history}
\usage{
get_anno_storage()

clear_anno_storage()
}
\value{
A list of grobs

A list of grobs
}
\description{
Get annotation history

Clear annotation history
}
