% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{grid_locate}
\alias{grid_locate}
\title{Interactively locate a point}
\usage{
grid_locate()
}
\value{
List of x and y positions in normalized parent coordinates, rounded to 3 digits
}
\description{
Interactively locate a point
}
