% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{qdraw}
\alias{qdraw}
\title{Quickly draw a grob or a list of grobs}
\usage{
qdraw(g, new = TRUE)
}
\arguments{
\item{g}{A grob or a list of grobs}

\item{new}{Whether to draw on a new page}
}
\value{
Current annotated figure as a grob
}
\description{
Quickly draw a grob or a list of grobs
}
