% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{undo_anno}
\alias{undo_anno}
\alias{redo_anno}
\title{Undo annotation}
\usage{
undo_anno()

redo_anno()
}
\value{
Current annotated figure as a grob

Current annotated figure as a grob
}
\description{
Undo annotation

Redo annotation
}
